+++ packages/playground/src/App.tsx
@@ -1,5 +1,8 @@
 import { FC } from "react"
 import { DateInput, HeroUIProvider } from "@heroui/react"
+import { pointInPolygon } from "soda-coordinate"
+
+type A = typeof pointInPolygon
 
 const App: FC = () => {
     return (