+++ packages/react-soda/.fatherrc.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from "father"
-
-export default defineConfig({
-    esm: {},
-    cjs: {},
-    sourcemap: true,
-})