+++ packages/soda-hooks/src/hooks/useSize.ts
@@ -1,5 +1,5 @@
-import { isNonNullable } from "deepsea-tools"
 import { RefObject, useEffect, useRef, useState } from "react"
+import { isNonNullable } from "deepsea-tools"
 
 export interface Size {
     width: number