+++ packages/soda-coordinate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-coordinate
 
+## 5.3.0
+
+### Minor Changes
+
+- 迁移到 rslib
+
 ## 5.2.9
 
 ### Patch Changes