+++ packages/soda-next/CHANGELOG.md
@@ -1,5 +1,17 @@
 # soda-next
 
+## 0.5.0
+
+### Minor Changes
+
+- 迁移到 rslib
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.25.0
+    - soda-hooks@6.7.0
+
 ## 0.4.42
 
 ### Patch Changes